pyproject.toml
[project]
name = "algo-trading"
version = "0.0.1"
dependencies = [
    "pandas~=2.2.3",
    "numpy~=1.26.4",
    "matplotlib~=3.8.4",
    "requests~=2.32.3",
    "schedule~=1.2.2",
    "MetaTrader5",
    "backtrader",
    "scipy",
    "mplfinance",
    "backtesting",
    "TA-Lib", # Install ta-lib binaries for this to work
    "python-dotenv"
]